refactor: move method into abstract handler
